检测图片

+++ routers/baseInfo.js
@@ -46,11 +46,13 @@ module.exports = (db) => {
                 if (logoUrl) {
                     let tmp_path = path.join(__dirname, `../assets${logoUrl}`);
                     if(tmp_path != defaultUrl){
-                        await fs.unlink(tmp_path, (err) => {
+                        if(fs.existssync(tmp_path)){
+                            await fs.unlink(tmp_path, (err) => {
                             if (err) {
                                 throw `error with unlink imageFile:${err}`;
                             }
-                        });
+                            });
+                        }
                     }
                 }
                 return ctx.body = {
